refactor(1396): split WASM worker bindings into per-category sub-packages
Fixes Cloudflare Worker TinyGo size failure in the prior Phase 2 commit:
registering all 55 bindings in a single games_wasm.go init() pinned every
game's domain/usecase/controller/presenter code in every worker binary,
which blew each one past the 1 MB gzipped free-tier limit.

- registry.go now holds bare metadata only ({Name, Category} for 55 games)
  plus BindWebController / BindWorker setters. No factory closures live
  at package scope, so TinyGo has nothing to anchor game code to.
- games_server.go (!js || !wasm) attaches Web-server factories via init()
  — excluded entirely from WASM builds.
- games/{casino,classic,solo}/<cat>.go (js && wasm) each bind only their
  18/21/16 games. Worker mains blank-import just one sub-package, so the
  other two categories' deps stay unreferenced and TinyGo DCE drops them.
- BindWorker cross-checks each binding's Category against the registry
  and panics on mismatch, turning sub-package mis-placements into
  package-load failures rather than silent routing bugs.
- Docs (CLAUDE.md, new-game-checklist.md) updated for the 3-touch
  flow: registry entry + Web bind + per-category Worker bind.
